// MSR.UI/MSR.UI.Answer/src/app/services/mocks/models/createProcedureTypeRequest.ts
/**
 * Parts, Procedures, and Monitors
 * REST API Answer 3
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class CreateProcedureTypeRequest { 
    name: string;
    majorGroup?: string;
}