// MSR.UI/MSR.UI.Answer/src/app/services/mocks/models/updateProcedureTemplateRequest.ts
/**
 * Parts, Procedures, and Monitors
 * REST API Answer 3
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Role } from './role';

export class UpdateProcedureTemplateRequest { 
    id?: number;
    title?: string;
    baseStartOnCounter?: boolean;
    estimatedStepDuration?: number;
    replacementCost?: number;
    utilization?: number;
    usefulLife?: number;
    text?: string;
    procedureStepId?: number;
    numberOfQuestionsToUse?: number;
    comments?: string;
    referenceProcedures?: Array<number>;
    referenceDocuments?: Array<number>;
    referenceFiles?: Array<any>;
    roles?: Array<number>;
    procedureStepTypeId:number;
}