// MSR.UI/MSR.UI.Answer/src/app/services/mocks/models/procedure.ts
/**
 * Parts, Procedures, and Monitors
 * REST API Answer 3
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProcedureTypeMock } from './procedureTypeMock';
import { Role } from './role';

export class Procedure { 
    id?: number;
    name?: string;
    isRelatedToAProduct?: boolean;
    creatorCompany?: string;
    createdByDepartmentName?: string;
    revision?: number;
    comment?: string;
    duration?: number;
    durationType?: string;
    procedureType?: ProcedureTypeMock;
    referenceFiles?: Array<any>;
    roles?: Array<Role>;
}